Removed the now unnecessary Swing form code
Related to #267
