--- src/main/java/fi/aalto/cs/apluscourses/ui/module/ModuleListElementView.form
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="fi.aalto.cs.apluscourses.ui.module.ModuleListElementView">
-  <grid id="27dc6" binding="basePanel" layout-manager="FlowLayout" hgap="5" vgap="5" flow-align="3">
-    <constraints>
-      <xy x="20" y="20" width="500" height="400"/>
-    </constraints>
-    <properties>
-      <toolTipText value=""/>
-    </properties>
-    <border type="none"/>
-    <children>
-      <component id="e26e5" class="javax.swing.JLabel" binding="nameLabel">
-        <constraints/>
-        <properties>
-          <text resource-bundle="resources" key="ui.toolWindow.subTab.modules.module.name"/>
-        </properties>
-      </component>
-      <component id="80135" class="javax.swing.JLabel" binding="statusLabel">
-        <constraints/>
-        <properties>
-          <font size="10" style="2"/>
-          <text resource-bundle="resources" key="ui.toolWindow.subTab.modules.module.status"/>
-        </properties>
-      </component>
-      <component id="c424d" class="javax.swing.JLabel" binding="updateLabel">
-        <constraints/>
-        <properties>
-          <font style="1"/>
-          <foreground color="-98404"/>
-          <text resource-bundle="resources" key="ui.toolWindow.subTab.modules.module.updateAvailable"/>
-        </properties>
-      </component>
-    </children>
-  </grid>
-</form>
